backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
pymongo==4.6.0
sqlalchemy==2.0.23
pandas==2.1.3
plotly==5.18.0
Pillow==10.1.0
ImageHash==4.3.1
requests==2.31.0
PyMySQL==1.1.0
psycopg2-binary==2.9.9
openai==1.3.0
anthropic==0.7.8
